Refactor animation toolbar to use mantine components
