+++ src/renderer/components/animator/Animator/Animator.tsx
@@ -1,12 +1,12 @@
 import Content from "../../common/Content/Content";
 import Page from "../../common/Page/Page";
 import PageBody from "../../common/PageBody/PageBody";
-import { AnimationToolbar } from "../AnimationToolbar/AnimationToolbar";
 import { PreviewToolbar } from "../PreviewToolbar/PreviewToolbar";
 import { CapturePane } from "../CapturePane/CapturePane";
 import { Preview } from "../Preview/Preview";
 import { Timeline } from "../Timeline/Timeline";
 import TitleToolbar from "../TitleToolbar/TitleToolbar";
+import { FrameToolbar } from "../FrameToolbar/FrameToolbar";
 
 export const Animator = (): JSX.Element => {
   return (
@@ -16,7 +16,7 @@ export const Animator = (): JSX.Element => {
         <Content>
           <Preview />
           <PreviewToolbar />
-          <AnimationToolbar />
+          <FrameToolbar />
           <Timeline />
         </Content>
